Move code that does work out of the configdomain package
