+++ internal/config/configdomain/branch_type_overrides.go
@@ -1,12 +1,7 @@
 package configdomain
 
 import (
-	"fmt"
-	"strings"
-
-	"github.com/git-town/git-town/v21/internal/cli/colors"
 	"github.com/git-town/git-town/v21/internal/git/gitdomain"
-	"github.com/git-town/git-town/v21/internal/messages"
 )
 
 // BranchTypeOverrides contains all configured branch type overrides.
@@ -23,37 +18,4 @@ func (self BranchTypeOverrides) Concat(other BranchTypeOverrides) BranchTypeOver
 		result[key] = value
 	}
 	return result
-}
-
-// provides the branch type overrides stored in the given Git metadata snapshot
-func NewBranchTypeOverridesInSnapshot(snapshot SingleSnapshot, gitCommands configRemover) (BranchTypeOverrides, error) {
-	result := BranchTypeOverrides{}
-	for key, value := range snapshot.BranchTypeOverrideEntries() {
-		branch := key.Branch()
-		if branch == "" {
-			// empty branch name --> delete it
-			fmt.Println(colors.Cyan().Styled(messages.ConfigBranchTypeOverrideEmpty))
-			_ = gitCommands.RemoveConfigValue(ConfigScopeLocal, key.Key)
-			continue
-		}
-		value = strings.TrimSpace(value)
-		if value == "" {
-			// empty branch type values are invalid --> delete it
-			fmt.Println(colors.Cyan().Styled(messages.ConfigBranchTypeOverrideEmpty))
-			_ = gitCommands.RemoveConfigValue(ConfigScopeLocal, key.Key)
-			continue
-		}
-		branchTypeOpt, err := ParseBranchType(value)
-		if err != nil {
-			return result, err
-		}
-		if branchType, hasBranchType := branchTypeOpt.Get(); hasBranchType {
-			result[branch] = branchType
-		}
-	}
-	return result, nil
-}
-
-type configRemover interface {
-	RemoveConfigValue(ConfigScope, Key) error
 }
