+++ internal/config/configdomain/branch_type_overrides_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/git-town/git-town/v21/internal/config/configdomain"
-	"github.com/git-town/git-town/v21/internal/config/gitconfig"
 	"github.com/shoenig/test/must"
 )
 
@@ -30,21 +29,4 @@ func TestBranchTypeOverrides(t *testing.T) {
 		must.Eq(t, want, have)
 	})
 
-	t.Run("NewBranchTypeOverridesFromSnapshot", func(t *testing.T) {
-		t.Parallel()
-		snapshot := configdomain.SingleSnapshot{
-			"git-town-branch.branch-1.branchtype": "feature",
-			"git-town-branch.branch-2.branchtype": "observed",
-			"git-town-branch.branch-3.parent":     "main",
-			"git-town.prototype-branches":         "foo",
-		}
-		gitIO := gitconfig.IO{Shell: nil}
-		have, err := configdomain.NewBranchTypeOverridesInSnapshot(snapshot, &gitIO)
-		must.NoError(t, err)
-		want := configdomain.BranchTypeOverrides{
-			"branch-1": configdomain.BranchTypeFeatureBranch,
-			"branch-2": configdomain.BranchTypeObservedBranch,
-		}
-		must.Eq(t, want, have)
-	})
 }
